lisa._kmod: Add util_est_unified trace events
FEATURE and FIX

Upstream has now removed struct util_est and effectively merged
util_est.enqueued and util_est.ewma in a single int field. Provide a new
trace event for it so the new kernel can trace it.

This also fixes LISA not being able to compile and load the module on
the new kernel. We detect whether struct util_est exists to decide if
the new signal should be used.