tools/exekall: Do not consider inherited methods defined outside of the package set
FIX

Restrict the operators considered by exekall to the ones defined in the
same package as the modules that exekall is pointed at. Also apply this
restriction to inherited methods, so we do not accidentally pull
operators from third party library that might use type annotations
differently and throw exekall into lengthy computations.