tools/trace-parser: Fix event selection
FIX

Event selection is done by comparing the event name with the vector of
selected names. That is supposed to be done once, and the result of that
decision is recorded in the StateMap. Before this commit, the selection
procedure was repeated every time for non-selected events, leading to a
waste of time.