lisa.trace: Preload trace metadata
FEATURE

Preload trace metadata from the cache so that all the side effects of
knowning a given piece of metadata has a chance to execute before the
user makes any query. For example, knowing the set of available events
allows rejecting very quickly requests for events that do not exist
without spinning up a parser.