tools/trace-parser: Do not emit event info key if event did not occur
FIX

If the an event does not appear in the trace, do not create an event
info key as this is used by the Python API to infer the available
events.

One day, trace.dat may record the trace-cmd CLI so we can infer what
events were enabled. If and when this happens, we can emit an empty
parquet file and event info when an enabled event did not occur, but for
now we cannot distinguish it from "event not enabled".