lisa.trace: Convert TxtTraceParserBase internals to polars
BREAKING CHANGE

Convert the internals of TxtTraceParserBase to polars.

This unfortunately comes at a loss of feature as polars' regexes (based
on the "regex" Rust crate at the time of writing) does not handle
look-around (look-behind and look-ahead). In turn, that prevents from
parsing correctly string fields with spaces in the values, e.g.

    comm=hello world anotherfield=42

This change is deemed acceptable as this fully removes the dependency on
pandas for lisa.trace (apart from converting legacy pandas inputs), and
the human-readable format of trace-cmd is ambiguous in any case. The
only ftrace format currently really supported is the binary trace.dat
format created by trace-cmd, which does not suffer from these issues.