lisa.analysis.base: Set plot height with output="ui"
FIX

Panel does not handle properly plots where "aspect" and
"responsive=True" has been set, as holoviews is apparently unable to
provide a plot size for panel, leading to the plot be squished.

Instead, specify height and responsive=True allows both systems to
interact as expected.

This changes the behavior as the aspect ratio is not preserved anymore
when the width of the jupyter lab window is reduced. That is however
probably an improvement as most (all ?) plots are using time for the X
axis, so reducing the width of the window just compresses time without
making the plot too small to be useful.