lisa.target: Resolve known_hosts path when strict_host_check=True
FIX

When strict_host_check=True, resolve the path to known_hosts SSH file so
that the target object can be reused without troubles in a different
user namespace. Otherwise, the home folder would resolve to another
path, leading to not being able to reload the file content and
ultimately failure of the check.