lisa._kmod: Dissociate Alpine chroot install and Rust install
FEATURE

Dissociate the creation of the chroot in the DirCache with the
installation of the Rust toolchain. This allows re-using an existing
chroot that was setup without Rust (e.g. to prepare the kernel tree) and
then simply add a Rust toolchain on top in a separate DirCache
bind-mounted in the alpine chroot. This way, we recycle the existing
chroot, rather than having to recreate it.