lisa.version: Make VERSION_TOKEN only depend on lisa sources
FIX

In dev mode, VERSION_TOKEN contains a hash of the versioned but
uncomitted content in the repo. This is far too wide, as this would
include things like target_conf.yml.

Instead, restrict the check to the lisa Python package sources, so that
we do not end up invalidating e.g. dataframe caches just because someone
changed target_conf.yml .