lisa._assets.kmodules.lisa: Support older linkers
FIX

Some older versions of LD seem to ignore an EXTERN() command in a linker
script when it comes to --gc-sections behavior, whereas it is suposed to
be the same as passing --undefined on the command line (which works).

Avoid using EXTERN() command and use the --undefined CLI parameter
instead.