lisa._kmod: Improve exception debug logging
If an exception is raised when selecting a kernel tree loader, print the
traceback in debug log if the error comes from the guts of the logic. If
this is a surface error intended for the user, do not print the
traceback as this would just mislead into thinking there is something
wrong with the logic.