lisa._kmod: Fixup Alpine chroot binary names
FIX

The binaries for the version of clang and LLVM that Alpine ships as its
default version will not have any version number in their name. This is
an issue as it is not trivial to know what version Alpine uses by
default, and there is no package providing versioned name either.

Fix that by creating shims for versioned names that simply redirect to
the normal binary when necessary.