lisa._assets.kmodules.lisa: Makefile: Fix grouped rule
FIX

Makefile rules with multiple targets is just a short-hand to duplicating
the recipe for each of those targets. Unfortunately, that means that if
make needs to build all these targets, it will run the same recipe
concurrently multiple times. This leads to file corruption as multiple
tools are writing to them concurrently.

Fix that by splitting the rule such that we do not need multiple targets
for the same rule anymore.