lisa.analysis._pixel: Add "energy" column to df_power_meter()
FEATURE

Convert to polars and add "energy" column since it is basically free on
a polars LazyFrame.

The code is also now aligned better with the documentation, as the power
is now computed over the period ending at the current row rather than
the period starting at the current row. This brings a null initial value
in the power column, for which there is an energy but no power
available.