lisa._kmod: Cleanly disconnect from Target in subprocess
FIX

When we run in a subprocess with separate user namespace in order to
gain root privileges, ensure we disconnect from the target when we are
done as this target is a copy (via pickle) of the one we got from the
parent process. Failing to disconnect can result in exceptions as
connections get closed upon garbage collection during interpreter
shutdown as part of the stdlib is not available anymore.