lisa.datautils: Remove df_refit_index(method=...) parameter
BREAKING CHANGE

Remove the "method" parameter and make the function effectively match
the behavior of passing method='pre'.

The parameter was never needed in practice and made implementing a
streaming version harder for no benefit. Changing the behavior from
"inclusive" to "pre" means the last row added by df_refit_index() will
not be reflecting the future value of the signal but instead the actual
current value of the signal at that timestamp. This should have very
little effect on real-world computations, as the last row is usually not
really used for anything else than its timestamp.