lisa: Ensure DirCache always has fmt_version set
FIX

Ensure each DirCache manages its own format versioning rather than
relying on lisa.version.VERSION_TOKEN, as the latter is unreliable in
the case of lisa being installed in non-editable mode from a
non-officially released version (e.g. installing the "main" version of
the day).

It works correctly when lisa has been installed in editable mode from a
git tree, but is otherwise equivalent to the hardcoded
lisa.version.__version__, which is only updated when doing an external
release.