lisa._kmod: Workaround new rustup behavior
FIX

rustup 1.28.0 started using absolute symlinks for some toolchain files
rather than using hardlinks. That is a problem for LISA as the Rust
toolchain is installed in a temporary folder that is then moved under a
long-term name. When the folder gets rename, all symlinks break as they
are absolute:

    https://github.com/rust-lang/rustup/issues/4222

Fix that by adding a post-processing pass after running rustup so that
symlinks get re-written as relative links.