lisa._kmod: Do not uninstall() kmod before installing it
BREAKING CHANGE

Changes DynamicKmod.install() and DynamicKmod.run() to not attempt to
uninstall the module before installing it. This will increase the
chances of install failing if the module was already loaded but is
necessary to not wipe existing module configuration in nested cases.

As a result, installing the module after having interrupted e.g. the
DynamicKmod.run() context manager may fail as the module will be loaded
already. It will then be on the user to cleanup the system and manually
rmmod the module.