install_base.sh: Remove --install-bisector-dbus
BREAKING CHANGE

Remove --install-bisector-dbus and the associated infrastructure in
install_base.sh as it is broken. Bisector is not used much anymore, and
its DBus support even less. DBus features also haven't been tested in
years, and given the difficulty of installing the associated libraries
when it was developped, the chances it has bit rotted are extremely
high.