lisa._assets.kmodules.lisa: Add equivalent of tracepoint_string()
FEATURE

Allow ftrace events fields to contain a reference to a
statically-allocated string that is stored in the event as a single
pointer, referencing a string stored in the header.

This allows a very compact encoding for strings at the cost of not being
able to dynamically be able to register new such strings.