lisa.trace: Reduce parquet cache hardlinks tracking overhead
FIX

LazyFrames loaded from the cache refer to a hardlink to the parquet
file. They own this hardlink and upon destruction of the LazyFrame, the
hardlink is removed. Since LazyFrames do not support attaching custom
metadata, we use LazyFrame.map_batches() with an identity callable and
we run hardlink freeing code in the destructor of that callable.

Unfortunately, this callable will be called for every chunk of data
loaded from the parquet file, and if the file has very few rows, this
function will be called with very small batches. This is a problem for
analysis functions that split the work e.g. per task, where the whole
dataset ends up being split across many small files. When all those
files are reloaded and e.g. concatenated, the overhead of the callable
is high.

In order to reduce this overhead, only track the hardlink for
large-enough dataframes. for smaller ones, simply collect immediately
and delete the hardlink straight away.