lisa.trace: Add delayed events parsing
FEATURE

When polars LazyFrame are used, we can delay event parsing under some
circumstances to the point where the LazyFrame schema is actually needed
(typically when calling LazyFrame.collect()).

This allows spinning up the parser only once to parse more than one
event at a time, which is typically much faster than parsing each event
dataframe one by one.

If the user's computation is organized such that calling
LazyFrame.collect() is the very last thing done, it will be almost as
efficient as using explicit event preloading. The only difference will
be that the parser needs to be queried for available events upfront,
which is not free. However, that cost is normally amortized if 2 or more
dataframe can be parsed from the trace at once.