lisa.analysis.rta: Silence pandas warning
FIX

Silence the following warning:

FutureWarning: DataFrameGroupBy.apply operated on the grouping columns.
This behavior is deprecated, and in a future version of pandas the
grouping columns will be excluded from the operation. Either pass
`include_groups=False` to exclude the groupings or explicitly select the
grouping columns after groupby to silence this warning.